separate frontend works
The JavaScript is very minimal, but enough to read the information in the fields, and send a request to the backend. Bokeh plots are not implemented, because I didn't want to send all that info over the network for a not-so-interesting graph.

If we need a visualization, can we do something with the pages text